card game made for game AI class
It has a metric truckload of issues, 90% sure it had a memory leak, looking at that nested if statement gives me nightmares (I should have known better)  and I have definitely grown as a coder over the past couple of years.  But it is cool so its got that going for it.